+++ src/test/java/net/dv8tion/jda/entitystring/ASnowflake.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.dv8tion.jda;
+package net.dv8tion.jda.entitystring;
 
 import net.dv8tion.jda.api.entities.ISnowflake;
